# examples/openapi.yaml
openapi: 3.1.0
info:
  title: Scribe Demo API
  version: 1.0.0
  summary: Backend contract for the Scribe metered-AI-transcription demo.
  description: |
    This is the contract for the **demo application's own backend** — not Meteroid's REST API.

    Scribe is a fictional metered AI transcription SaaS used to demonstrate the Meteroid SDKs.
    Every backend under `examples/backends/<language>/` implements this contract by hand using the
    Meteroid SDK for that language, and the single SPA under `examples/frontend/` consumes it.
    The contract suite under `examples/tests/contract/` runs against any `BASE_URL` and validates
    every response against the schemas below, so conformance is enforced here rather than by a
    code generator.

    ## Modelling rules this contract follows

    * **Decimals are strings.** Everything Meteroid serializes as `format: decimal` (usage totals,
      entitlement limits, consumed/remaining) stays a string here. Never a JSON number — the values
      are exact decimals and must survive a round trip.
    * **Money on invoices is an integer in minor units.** Meteroid's invoice amounts are `int64`
      minor units (cents), *not* decimals; this contract keeps them integers and says so on each field.
    * **Unions are explicitly tagged.** Entitlement values and config values are discriminated by a
      literal `type` / `kind` property, mirroring Meteroid's own discriminators. No untagged `oneOf`.
...

    ## Authentication

...

...

...

...

...

  license:
...
servers:
...
security:
  - sessionToken: []
paths:
...
  /api/session:
    post:
      operationId: createSession
      tags: [session]
      summary: Create a demo workspace (and a Meteroid customer).
      description: |
        Creates a brand-new Meteroid customer with a generated, unique alias and returns a session
        token bound to it. Every demo run gets its own customer; the same Meteroid API key is reused
        indefinitely. No subscription exists yet — the workspace starts unsubscribed, and the
        frontend drives the visitor to `POST /api/checkout`.

        Meteroid's `CustomerCreateRequest` requires `name`, `invoicing_emails`, `currency` and
        `custom_taxes`. The backend fills them from the (optional) request body plus generated
        defaults, and takes the currency from `SCRIBE_DEFAULT_CURRENCY`. **That currency must equal
        the currency of the seeded plans** — Meteroid rejects a checkout whose customer and plan
        version disagree — so a mismatch is reported as `CATALOG_NOT_SEEDED` at checkout time.

        The request body is optional: no body, an empty body and `{}` are all equivalent.

        **SDK surface:** Rust `client.customers().create_customer(CustomerCreateRequest { .. })`;
        Java `new Customers(httpClient).createCustomer(request)`.
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Workspace created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        # Deliberately no 503: creating a customer touches no catalog object. The
        # currency mismatch described above is detected at `POST /api/checkout`, which
        # is the first operation that has to reconcile the customer with a plan version.
  /api/me:
    get:
      operationId: getMe
      tags: [session]
      summary: The current workspace, its subscription and its plan.
      description: |
        Resolves the session's customer, its most relevant subscription, and the plan that
        subscription is on. `subscription` and `plan` are both `null` for a workspace that has
        never checked out.

        Subscriptions are fetched with `GET /api/v1/subscriptions?customer_id=<alias>`, which accepts
        a customer id *or* alias, so no extra id lookup is needed.

        **Which subscription is "the most relevant" one** is pinned down here so every backend
        picks the same one, and so `GET /api/usage` and `POST /api/transcriptions` agree with this
        endpoint: ask Meteroid with `order_by=created_at.desc`, take the first subscription whose
        status is `ACTIVE` or `TRIAL_ACTIVE`, and otherwise the first of the list (the most
        recently created, whatever its status).

        The seeded catalog is consulted only when a subscription exists — it is what maps
        Meteroid's `plan_id`/`plan_name` onto this contract's `plan_code`. A workspace that has
        never subscribed therefore answers even on a tenant whose catalog is unreachable, while a
        subscribed one reports `CATALOG_NOT_SEEDED`; that is deliberate, because this is the
        frontend's boot call and an unseeded tenant should be visible immediately rather than
        first surfacing at checkout.

        **SDK surface:** Rust `client.customers().get_customer(alias)` and
        `client.subscriptions().list_subscriptions(..)`; Java `getCustomer` / `listSubscriptions`.
      responses:
        "200":
          description: Current workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        "503": { $ref: "#/components/responses/CatalogNotSeeded" }
  /api/plans:
    get:
      operationId: listPlans
      tags: [catalog]
      summary: The pricing table.
      description: |
        The three seeded Scribe plans, each resolved to its latest published (non-draft) version,
        with a flattened summary of that version's price components. Ordered `free`, `pro`, `scale`.

        Meteroid plans have no user-supplied `code`, so the backend resolves each plan by its exact
        seeded **name** (see `examples/CATALOG.md`) and maps it onto the stable `plan_code` this
        contract uses. `GET /api/v1/plans?search=…` is a fuzzy name match, so the backend must
        additionally require `name == <seeded name>` exactly. A missing or unpublished plan is a
        seeding error, not a runtime condition: the backend returns `CATALOG_NOT_SEEDED` naming the
        plan it could not resolve.

        `GET /api/v1/plans` already returns full `Plan` objects, including `version_id` and
        `price_components`, so the pricing table needs no per-plan follow-up call; the marketing
        bullets in `features` come from `GET /api/v1/plan-versions/{version_id}/entitlements`, whose
        `ResolvedEntitlement.feature` carries the feature `code`.

        **SDK surface:** Rust `client.plans().list_plans(..)` and
        `client.plans().list_plan_version_entitlements(plan_version_id)`; Java `listPlans` /
        `listPlanVersionEntitlements`.
      security: []
      responses:
        "200":
          description: The pricing table.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanListResponse"
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        "503": { $ref: "#/components/responses/CatalogNotSeeded" }
  /api/checkout:
    post:
      operationId: createCheckout
      tags: [billing]
      summary: Start a hosted checkout for a plan.
      description: |
        Creates a Meteroid checkout session for the workspace's customer against the requested
        plan's published version and returns the hosted checkout URL to redirect to. Used both for
        the first subscription and for upgrades — Meteroid decides the `checkout_type`.

        A Meteroid checkout session may legitimately carry a null `checkout_url` (non-self-serve
        checkout types). This contract treats that as a hard failure for the demo's purposes and
        returns `CHECKOUT_UNAVAILABLE` rather than handing the frontend a null it cannot use.

        Meteroid's `CreateCheckoutSessionRequest.customer_id` accepts a customer id *or* alias, so
        the backend passes the workspace alias straight through.

        **SDK surface:** Rust `client.checkout_sessions().create_checkout_session(req)`;
        Java `createCheckoutSession(request)`. The response is `{ session: CheckoutSession }`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutRequest"
      responses:
        "201":
          description: Checkout session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCheckoutResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: |
            Checkout cannot be started for this workspace and plan — for example the workspace is
            already subscribed to the requested plan version.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                unavailable:
                  value:
                    code: CHECKOUT_UNAVAILABLE
                    message: "Meteroid returned a checkout session without a hosted URL (checkout_type=PLAN_CHANGE)."
                    quota: null
                    upgrade_plan_code: null
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        "503": { $ref: "#/components/responses/CatalogNotSeeded" }
  /api/entitlements:
    get:
      operationId: listEntitlements
      tags: [entitlements]
      summary: Normalized entitlements for feature gating.
      description: |
        The workspace's effective entitlements, one entry per feature, normalized into an explicitly
        tagged union the frontend can switch on. This is the view the SPA gates the UI with and the
        same view `POST /api/transcriptions` enforces server-side.

        Meteroid's `EffectiveEntitlementValue` is a `oneOf` discriminated on `type`
        (`BOOLEAN` / `METERED` / `CONFIG`); this contract keeps that discriminator, flattens the
        metered `spec`+`usage` pair into one object, and preserves every decimal as a string.

        Note that Meteroid's `MeteredEntitlementUsage` declares *no* required properties, so
        `consumed`, `remaining` and `reset_at` may be absent on the wire. Backends normalize an
        absent value to an explicit `null` here, per this contract's response nullability rule.

        Expected feature codes for the seeded Scribe catalog: `transcription_minutes` (metered),
        `sso` (boolean), `retention_days` (config/number), `seats` (config/number). A workspace with
        no subscription still returns the feature-level defaults Meteroid resolves, which may be an
        empty list.

        **SDK surface:** Rust `client.customers().get_effective_entitlements(alias)`;
        Java `getEffectiveEntitlements(idOrAlias)`. The response is `{ data: EffectiveEntitlement[] }`.
      responses:
        "200":
          description: Effective entitlements.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntitlementListResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
        "502": { $ref: "#/components/responses/UpstreamError" }
  /api/transcriptions:
    get:
      operationId: listTranscriptions
      tags: [usage]
      summary: Transcriptions created in this workspace.
      description: |
        Demo-local history, newest first. Backends keep this in memory; it is not persisted and it
        resets when the process restarts. Meteroid is the source of truth for *usage*, not for the
        application objects that produced it.
      responses:
        "200":
          description: Transcription history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionListResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }

    post:
      operationId: createTranscription
      tags: [usage]
      summary: Transcribe audio — the metered action.
      description: |
        The centerpiece of the demo. In order:

        1. Read the workspace's effective entitlements and find `transcription_minutes`.
        2. If the entitlement is missing or `enabled` is `false` → **403 `FEATURE_NOT_ENTITLED`**.
           Note that "missing" is ambiguous per-request: an unsubscribed workspace and an unseeded
           tenant both produce an empty entitlement list. Backends resolve the ambiguity once at
           startup with `GET /api/v1/features/{code}` for the four seeded codes, and report an
           unseeded tenant as `CATALOG_NOT_SEEDED` rather than letting it masquerade as a 403.

           If the entitlement is *present but not `METERED`* — the feature exists in the tenant but
           was created as boolean or config in the dashboard — that is a seeding mistake, not a
           customer-facing paywall, and it is **503 `CATALOG_NOT_SEEDED`**. Never a 403: an
           operator error must not look like "your plan doesn't include that".
        3. If it has a `limit` and the requested minutes exceed the remaining balance →
           **402 `QUOTA_EXHAUSTED`**, with the live quota snapshot and the plan to upgrade to
           echoed back in the error body so the frontend can render an upgrade prompt without a
           second round trip. A `null` limit means unlimited and never trips this branch.

           The remaining balance is Meteroid's `remaining`; because Meteroid may supply a `limit`
           with no counter yet, backends fall back to `limit - consumed` when `remaining` is null,
           and to `limit` when both are null. All three are decimal strings and the arithmetic must
           be exact decimal arithmetic (`rust_decimal`, `BigDecimal`) — never binary floating point,
           which is how a 0.1-minute clip eventually bills wrong.
        4. Otherwise produce the (fake) transcript and report consumption to Meteroid by ingesting
           one event with `code: transcription_minutes`, the workspace's customer alias, and
           `properties.minutes` set to the billed minutes as a decimal string.

        Because step 2 turns a missing or disabled entitlement into a 403, a 201 always carries a
        `quota` — an unlimited entitlement is reported as `unlimited: true` with `limit`,
        `remaining` and `reset_at` null, never as an absent quota.

        Meteroid usage counters are eventually consistent: `quota` in the 201 body is the backend's
        optimistic local projection of the entitlement it just read, and `GET /api/usage` may lag by
        a few seconds. Callers must not treat the two as transactionally consistent.

        The ingest is a batch call (`{ events: [ … ] }`, 1–100 events); the demo sends exactly one,
        with `timestamp` set to now — Meteroid rejects timestamps outside `24h ago .. 1h ahead`
        unless `allow_backfilling` is set, which this demo never does.

        **SDK surface:** Rust `client.customers().get_effective_entitlements(alias)` then
        `client.events().ingest_events(IngestEventsRequest { events, .. })`;
        Java `getEffectiveEntitlements` then `ingestEvents`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTranscriptionRequest"
      responses:
        "201":
          description: Transcribed, and the consumed minutes were reported to Meteroid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTranscriptionResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { $ref: "#/components/responses/QuotaExhausted" }
        "403": { $ref: "#/components/responses/FeatureNotEntitled" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        "503": { $ref: "#/components/responses/CatalogNotSeeded" }
  /api/usage:
    get:
      operationId: getUsage
      tags: [usage]
      summary: Current-period usage for the workspace.
      description: |
        Usage for the current billing period, per billable metric, straight from Meteroid. Queried
        by subscription when the workspace has one (so the period matches the billing cycle),
        falling back to the customer-scoped query otherwise. `scope` says which happened.

        The two Meteroid endpoints differ in an important way:

        * `GET /api/v1/usage/subscription/{id}` takes `start_date`/`end_date` as **optional** query
          params; omitting them makes Meteroid use the subscription's current billing period. The
          backend omits them, so `period_start`/`period_end` are Meteroid's, not the demo's.
        * `GET /api/v1/usage/customer/{id_or_alias}` **requires** `start_date` and `end_date`. With
          no subscription there is no billing period to borrow, so the backend supplies the current
          calendar month (first of the month .. today, UTC) and echoes exactly that back. Treat a
          `customer`-scoped period as a demo convention, not a billing period.

        `total_value` and every grouped `value` are decimal strings.

        **SDK surface:** Rust `client.usage().get_subscription_usage(..)` /
        `client.usage().get_customer_usage(..)`; Java `getSubscriptionUsage` / `getCustomerUsage`.
      responses:
        "200":
          description: Usage for the current period.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
        "502": { $ref: "#/components/responses/UpstreamError" }
  /api/portal-session:
    post:
      operationId: createPortalSession
      tags: [billing]
      summary: Mint a customer-portal token.
      description: |
        Issues a short-lived Meteroid customer-portal token for the workspace's customer, so the
        visitor can manage their payment method and download invoices without the demo implementing
        any of it. The frontend opens `portal_url` with the `token`.

        Meteroid documents the valid lifetime as 60..2592000 seconds but types the field only as a
        non-negative int32, so the backend validates the range itself and returns `BAD_REQUEST`
        rather than forwarding a value Meteroid would reject. The request body is optional: no body,
        an empty body and `{}` all mean "use Meteroid's default of 86400 seconds".

        **SDK surface:** Rust `client.customers().create_portal_token(alias, req)`;
        Java `createPortalToken(idOrAlias, request)`. The response is `{ token, portal_url }` —
        Meteroid returns no expiry, so `expires_in_seconds` below is the value the backend asked for.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortalSessionRequest"
      responses:
        "201":
          description: Portal token minted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePortalSessionResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
        "502": { $ref: "#/components/responses/UpstreamError" }
  /api/invoices:
    get:
      operationId: listInvoices
      tags: [billing]
      summary: Invoices for the workspace.
      description: |
        The workspace customer's invoices, newest first. A fresh workspace has none, and a workspace
        that just checked out usually has a `DRAFT` one.

        Amounts are integers in the currency's **minor units** (cents), matching Meteroid.

        Fetched with `GET /api/v1/invoices?customer_id=<alias>&order_by=invoice_date.desc`;
        `customer_id` accepts an id or an alias, and `limit` below maps onto Meteroid's `per_page`
        (which also caps at 100).

        **SDK surface:** Rust `client.invoices().list_invoices(..)`; Java `listInvoices(options)`.
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum number of invoices to return.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Invoices.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
        "502": { $ref: "#/components/responses/UpstreamError" }
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        The `session_token` returned by `POST /api/session`. Opaque to clients. Not a Meteroid
        credential — the Meteroid API key never leaves the backend.
  responses:
    BadRequest:
      description: The request body or query is invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            invalid:
              value:
                code: BAD_REQUEST
                message: "duration_seconds must be between 1 and 7200."
                quota: null
                upgrade_plan_code: null
...
    RateLimited:
      description: |
        Meteroid throttled the backend (it returns `429` on every endpoint this demo calls) or the
        backend is throttling itself. Retry after a short backoff; the demo does not add a
        `Retry-After` header.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            throttled:
              value:
                code: RATE_LIMITED
                message: "Meteroid responded 429. Retry shortly."
                quota: null
                upgrade_plan_code: null
    InternalError:
      description: |
        An unhandled failure in the demo backend itself — a bug, not a Meteroid or configuration
        problem. Declared on every operation so the contract suite validates the envelope instead of
        hitting an untyped framework error page.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            internal:
              value:
                code: INTERNAL
                message: "Unexpected error handling the request."
                quota: null
                upgrade_plan_code: null
...
    UpstreamError:
      description: |
        The Meteroid API returned an error or was unreachable. `UPSTREAM_UNAUTHORIZED` specifically
        means the backend's `METEROID_API_KEY` is missing, wrong, or revoked — an operator problem.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            upstream:
              value:
                code: UPSTREAM_ERROR
                message: "Meteroid responded 500 to POST /api/v1/events/ingest."
                quota: null
                upgrade_plan_code: null
            bad_key:
              value:
                code: UPSTREAM_UNAUTHORIZED
                message: "Meteroid rejected the API key. Check METEROID_API_KEY."
                quota: null
                upgrade_plan_code: null
...
  schemas:
...
    CreateSessionRequest:
      type: object
      description: |
        All fields optional; the backend generates whatever is not supplied. Omitting a key, sending
        it as `null`, and sending no body at all are equivalent.
      additionalProperties: false
      properties:
        workspace_name:
          description: Display name for the workspace, also used as the Meteroid customer name.
          type: [string, "null"]
          minLength: 1
          maxLength: 120
        email:
          description: |
            Invoicing email for the Meteroid customer — Meteroid requires a non-empty
            `invoicing_emails` array, so the backend generates a demo address when this is omitted.
          type: [string, "null"]
          format: email
          maxLength: 254
...
    CreateSessionResponse:
      type: object
      required: [session_token, workspace]
      additionalProperties: false
      properties:
        session_token:
          type: string
          description: "Send as `Authorization: Bearer <session_token>`. Opaque."
          minLength: 1
        workspace:
          $ref: "#/components/schemas/Workspace"
...